Refactor Schools::UserInvite to be nested under Schools::DfeSignInApi
